pyproject.toml
[tool.poetry]
name = "anvil-cli"
version = "0.1.0"
description = "A powerful CLI tool for creative development workflows"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "anvil"}]
license = "MIT"
homepage = "https://github.com/yourusername/anvil-cli"
repository = "https://github.com/yourusername/anvil-cli"
keywords = ["cli", "development", "creative", "tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.7.0"
pillow = "^10.0.0"
rich = "^13.0.0"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
ruff = "^0.1.0"
mypy = "^1.0.0"
types-pillow = "^10.0.0"

[tool.poetry.scripts]
anvil = "anvil.cli:app"
anvil-shell = "anvil.repl:repl"

[tool.poetry.plugins."anvil.plugins"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "C90", "I", "N", "UP", "S", "B", "A", "C4", "DTZ", "T10", "EM", "G", "PIE", "T20", "Q", "RET", "SIM", "ARG", "PTH", "PGH", "RUF"]
ignore = ["S101", "S603", "S607"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "ARG001", "ARG002"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=anvil --cov-report=term-missing --cov-report=html --cov-fail-under=90" 